+++ ios/Sources/AppFeature/AppTab.swift
@@ -1,0 +1,74 @@
+import AboutFeature
+import ComposableArchitecture
+import FavoritesFeature
+import HomeFeature
+import MediaFeature
+import Styleguide
+import SwiftUI
+
+enum AppTab: CaseIterable {
+    case home
+    case media
+    case favorites
+    case about
+
+    var title: String {
+        switch self {
+        case .home:
+            return L10n.HomeScreen.title
+        case .media:
+            return L10n.MediaScreen.title
+        case .favorites:
+            return L10n.FavoriteScreen.title
+        case .about:
+            return L10n.AboutScreen.title
+        }
+    }
+
+    var image: Image {
+        switch self {
+        case .home:
+            return AssetImage.iconHome.image
+        case .media:
+            return AssetImage.iconBlog.image
+        case .favorites:
+            return AssetImage.iconStar.image
+        case .about:
+            return AssetImage.iconAbout.image
+        }
+    }
+
+    @ViewBuilder
+    func view(_ store: Store<AppTabState, AppTabAction>) -> some View {
+        switch self {
+        case .home:
+            HomeScreen(
+                store: store.scope(
+                    state: \.homeState,
+                    action: AppTabAction.home
+                )
+            )
+        case .media:
+            MediaScreen(
+                store: store.scope(
+                    state: \.mediaState,
+                    action: AppTabAction.media
+                )
+            )
+        case .favorites:
+            FavoritesScreen(
+                store: store.scope(
+                    state: \.favoritesState,
+                    action: AppTabAction.favorites
+                )
+            )
+        case .about:
+            AboutScreen(
+                store: store.scope(
+                    state: \.aboutState,
+                    action: AppTabAction.about
+                )
+            )
+        }
+    }
+}
