+++ ios/Sources/HomeFeature/HomeFeature.swift
@@ -3,20 +3,30 @@ import ComposableArchitecture
 import Model
 import Repository
 
-public enum HomeState: Equatable {
-    case needToInitialize
-    case initialized(HomeListState)
+public struct HomeState: Equatable {
+    let feedContents: [FeedContent]
+    public var message: String {
+        "Finished! 🤖"
+    }
 
-    public init() {
-        self = .needToInitialize
+    public var topic: FeedContent? {
+        feedContents.first
+    }
+
+    public var listFeedContents: [FeedContent] {
+        Array(feedContents.dropFirst())
+    }
+
+    public init(feedContents: [FeedContent]) {
+        self.feedContents = feedContents
     }
 }
 
 public enum HomeAction {
-    case refresh
-    case refreshResponse(Result<[FeedContent], KotlinError>)
-    case needRefresh
-    case homeList(HomeListAction)
+    case selectFeedContent
+    case tapFavorite(isFavorited: Bool, id: String)
+    case favoriteResponse(Result<String, KotlinError>)
+    case answerQuestionnaire
 }
 
 public struct HomeEnvironment {
@@ -29,34 +39,27 @@ public struct HomeEnvironment {
     }
 }
 
-public let homeReducer = Reducer<HomeState, HomeAction, HomeEnvironment>.combine(
-    homeListReducer.pullback(
-        state: /HomeState.initialized,
-        action: /HomeAction.homeList,
-        environment: {
-            .init(feedRepository: $0.feedRepository)
-        }
-    ),
-    .init { state, action, environment in
-        switch action {
-        case .refresh:
-            return environment.feedRepository.feedContents()
-                .catchToEffect()
-                .map(HomeAction.refreshResponse)
-        case let .refreshResponse(.success(feedContents)):
-            if !feedContents.isEmpty {
-                state = .initialized(.init(feedContents: feedContents))
-            }
-            return .none
-        case let .refreshResponse(.failure(error)):
-            print(error.localizedDescription)
-            // TODO: Error handling
-            return .none
-        case .needRefresh:
-            state = .needToInitialize
-            return .none
-        case .homeList:
-            return .none
-        }
-    }
-)
+public let homeReducer = Reducer<HomeState, HomeAction, HomeEnvironment> { _, _, _ in
+    return .none
+//        switch action {
+//        case .refresh:
+//            return environment.feedRepository.feedContents()
+//                .catchToEffect()
+//                .map(HomeAction.refreshResponse)
+//        case let .refreshResponse(.success(feedContents)):
+//            if !feedContents.isEmpty {
+//                state = .initialized(.init(feedContents: feedContents))
+//            }
+//            return .none
+//        case let .refreshResponse(.failure(error)):
+//            print(error.localizedDescription)
+//            // TODO: Error handling
+//            return .none
+//        case .needRefresh:
+//            state = .needToInitialize
+//            return .none
+//        case .homeList:
+//            return .none
+//        }
+//    }
+}
